Refactor UI into modular pages and widgets for operation sheet, logs, and settings
