+++ emapi-demo/flutter/lib/src/entities/emapi_demo_log_entry.dart
@@ -1,0 +1,13 @@
+import 'dart:typed_data';
+
+class EmapiDemoLogEntry {
+  const EmapiDemoLogEntry({
+    required this.title,
+    required this.message,
+    this.bytes,
+  });
+
+  final String title;
+  final String message;
+  final Uint8List? bytes;
+}
